+++ HISTORY.rst
@@ -33,6 +33,7 @@ Unreleased
 * ALCN-1018: Display quants moved in the stock.move form view
 * ALCN-1014: The user must first start the picking to validate after
 * ALCN-1012: Add filters on round instances and round templates
+* ALCN-1011: When the delivery address is used on a picking, search the itinerary with the company
 
 **Bugfixes**